Quellcode/requirements.txt
requests
beautifulsoup4
selenium
pandas
xlrd
openpyxl
webdriver-manager
python-dotenv